Finish implementing the new cameca class
The old camera is being deprecated and a new camera from earnopengl.com
is now being made the new camera for the Engine.

This commit also brings new callbacks to the tamble, replacing the old
GLFW_keyboardCallback.cpp file and functions. The intention now is to
improve the camera with new modes to compply with the Game Jam